+++ src/main/java/teymi15/kassistant/repository/RecipeRepository.java
@@ -22,7 +22,7 @@ import java.util.List;
  * The repository for recipe
  */
 @Transactional
-public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
+public interface RecipeRepository extends CrudRepository<Recipe, Long> {
 
     /**
      * get the list of all recipe
